+++ include/sk/io/ByteArrayInputStream.h
@@ -31,7 +31,7 @@ namespace sk {
         int read(char* buffer, int offset, int length);
         using AbstractInputStream::read;
         void close();
-        long long available() const;
+        uint64_t available() const;
         bool markSupported() const;
         int skip(int number);
         void mark(int readlimit);
