+++ lib/sk/io/File.cc
@@ -157,18 +157,18 @@ numericMode(const sk::util::String& mode)
   throw sk::io::IOException("Bad mode " + mode.inspect() + " in open()");
 }
 
-long long
+uint64_t
 sk::io::File::
 position() const
 {
-  long long n = lseek(getFileDescriptor().getFileNumber(), 0, SEEK_CUR);
+  uint64_t n = lseek(getFileDescriptor().getFileNumber(), 0, SEEK_CUR);
   if(n == -1) {
     throw sk::io::IOException(getName() + ": lseek() failed");
   }
   return n;
 }
 
-long long
+uint64_t
 sk::io::File::
 size() const
 {
